✨ improve stack management + ABI compliance
it wasn't found worth it to map the stack area on each call.
it is now mapped on the constructor, the size is passed there.
the default size is now 16MiB.

also, instead of passing a 'stack' argument to call(), the
user calls reserve() to reserve space on the stack, which
is returned in the form of a UnicornIO just like mem()
would. it's actually a context manager that has to be released
in strict reverse order because... well, it's a stack.

it's much easier to use; you can easily get the address of the
buffer, you don't have to handle alignment yourself, the
allocation can persist across calls, you can use read_struct
/ write_struct and read_str / write_str just like on any other
memory.

example:

    with emu.reserve(16, align=1) as buf, emu.reserve(8) as size:
        buf.write_str('hello')
        emu.call('process_buffer', buf.addr, size.addr)
        return size.read_struct('<Q')

now that call() doesn't have to handle mapping/unmapping stack
and reading/writing user stack, it gets much simpler. using
the reserve() call makes it even simpler.

call() can now handle more than 6 arguments by passing them on
the stack, and takes care to align end of parameters to 16 bytes
as ABI mandates.

call() also does things a bit more safely now, in regard to
cleaning up hooks, memory, etc.